pyproject.toml
[project]
name = "reca11-memory"
version = "0.1.0"
description = "Memory layer and SDK for LLMs"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.98.0",
    "pinecone>=7.3.0",
    "pymongo[srv]>=4.13.2",
    "python-dotenv>=1.1.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.35.0",
    "requests>=2.0.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["reca11"]